downloader/pyproject.toml
[tool.poetry]
name = "youtube-downloader"
version = "0.1.0"
description = "Service to download YouTube audio"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "downloader", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pytubefix = "^8.12.3"
google-cloud-storage = "^2.10.0"
pydantic = "^2.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"